refactor(middleware): move transpiler into middleware
Closes: #311
