+++ packages/middleware/package.json
@@ -5,12 +5,15 @@
   "keywords": ["hops", "webpack", "express"],
   "license": "MIT",
   "main": "index.js",
-  "files": ["index.js"],
+  "files": ["index.js", "transpile.js"],
   "repository": {
     "type": "git",
     "url": "https://github.com/xing/hops.git"
   },
   "dependencies": {
-    "hops-transpiler": "9.4.1"
+    "memory-fs": "^0.4.1",
+    "require-from-string": "^2.0.1",
+    "source-map-support": "^0.5.0",
+    "webpack": "^3.6.0"
   }
 }
