+++ packages/spec/package.json
@@ -17,7 +17,6 @@
     "hops-redux": "9.3.0",
     "hops-renderer": "9.4.1",
     "hops-server": "9.3.0",
-    "hops-transpiler": "9.4.1",
     "isomorphic-fetch": "^2.2.1",
     "jest": "^21.2.1",
     "jest-preset-hops": "9.3.2",
