--- packages/transpiler/CHANGELOG.md
@@ -1,44 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-<a name="9.4.1"></a>
-## [9.4.1](https://github.com/xing/hops/compare/v9.4.0...v9.4.1) (2018-01-08)
-
-
-
-
-**Note:** Version bump only for package hops-transpiler
-
-<a name="9.1.1"></a>
-## [9.1.1](https://github.com/xing/hops/compare/v9.1.0...v9.1.1) (2017-12-05)
-
-
-
-
-**Note:** Version bump only for package hops-transpiler
-
-<a name="7.2.0"></a>
-# [7.2.0](https://github.com/xing/hops/compare/v7.1.0...v7.2.0) (2017-10-18)
-
-
-
-
-**Note:** Version bump only for package hops-transpiler
-
-<a name="7.1.0"></a>
-# [7.1.0](https://github.com/xing/hops/compare/v7.0.0...v7.1.0) (2017-10-16)
-
-
-
-
-**Note:** Version bump only for package hops-transpiler
-
-<a name="7.0.0"></a>
-# [7.0.0](https://github.com/xing/hops/compare/v6.2.8...v7.0.0) (2017-10-13)
-
-
-### Features
-
-* **transpiler:** split hops packages, introduce changelog
