--- packages/transpiler/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "hops-transpiler",
-  "version": "9.4.1",
-  "description": "Hops wrapper for Webpack",
-  "keywords": ["hops", "webpack"],
-  "license": "MIT",
-  "main": "index.js",
-  "files": ["index.js"],
-  "dependencies": {
-    "memory-fs": "^0.4.1",
-    "require-from-string": "^2.0.1",
-    "source-map-support": "^0.5.0",
-    "webpack": "^3.6.0"
-  }
-}
